resolved files structure
