+++ src/02-dom/cards.html
@@ -64,21 +64,16 @@
 				</div>
 				<div id="box-1" class="card">
 					<p class="content">Card-1</p>
-					<button class="btn-add">Add Before</button>
-					<button class="btn-after">Add After</button><br />
-					<button class="btn-del">Delete</button>
-				</div>
-				<div id="box-2" class="card">
-					<p class="content">Card-2</p>
-					<button class="btn-add">Add Before</button>
-					<button class="btn-after">Add After</button><br />
-					<button class="btn-del">Delete</button>
+					<button id="btn-add">Add Before</button>
+					<button id="btn-after">Add After</button><br />
+					<button id="btn-del">Delete</button>
 				</div>
 			</div>
 			<div id="right">
 				<h1>Right Side</h1>
 			</div>
 		</div>
+		<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.1.0/jquery.min.js"></script>
 
 		<script src="./scripts/cards_main.js" type="module" async defer></script>
 		<!-- Optional JavaScript -->
