+++ src/02-dom/index.html
@@ -82,7 +82,7 @@
 			</p>
 		<![endif]-->
 
-		<script src="./scripts/main.js" type="module" async defer></script>
+		<script src="./scripts/functions_main.js" type="module" async defer></script>
 		<!-- Optional JavaScript -->
 		<!-- jQuery first, then Popper.js, then Bootstrap JS -->
 		<script
