--- src/02-dom/sandbox/index_cards.css
@@ -1,49 +1,0 @@
-@import url("https://fonts.googleapis.com/css?family=Ubuntu&display=swap");
-
-.body-class-cards {
-	font-family: "Ubuntu", sans-serif;
-}
-.box-1 {
-	background: lightblue;
-	grid-row: 2/3;
-	margin: 10px 5px 5px 10px;
-	text-align: center;
-}
-
-.box-2 {
-	background: lightcoral;
-	margin: 10px 5px 5px 10px;
-	text-align: center;
-}
-.box-3 {
-	background: lightseagreen;
-	margin: 10px 5px 5px 10px;
-	text-align: center;
-}
-.box-4 {
-	background: lightgreen;
-	margin: 10px 5px 10px 10px;
-	text-align: center;
-}
-.box-5 {
-	background: lightslategray;
-	grid-row: 1/6;
-	margin: 10px 10px 10px 5px;
-	text-align: center;
-}
-.box-6 {
-	text-align: center;
-	margin: auto;
-}
-main {
-	height: 100vh;
-	display: grid;
-	grid-template-columns: 1fr 1fr;
-	grid-template-rows: auto;
-	border: 5px solid gray;
-	margin: auto;
-	width: 600px;
-}
-h1 {
-	text-align: center;
-}
