--- src/02-dom/sandbox/index_cards.html
@@ -1,111 +1,0 @@
-<!DOCTYPE html>
-<!--[if lt IE 7]>      <html class="no-js lt-ie9 lt-ie8 lt-ie7"> <![endif]-->
-<!--[if IE 7]>         <html class="no-js lt-ie9 lt-ie8"> <![endif]-->
-<!--[if IE 8]>         <html class="no-js lt-ie9"> <![endif]-->
-<!--[if gt IE 8]><!-->
-<html class="no-js">
-	<!--<![endif]-->
-	<head>
-		<meta charset="utf-8" />
-		<meta http-equiv="X-UA-Compatible" content="IE=edge" />
-		<title>Cards</title>
-		<meta name="description" content="" />
-		<meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no" />
-		<!-- Bootstrap CSS -->
-		<link
-			rel="stylesheet"
-			href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css"
-			integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm"
-			crossorigin="anonymous"
-		/>
-		<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css" />
-		<link href="./index_cards.css" rel="stylesheet" type="text/css" />
-	</head>
-	<body class="body-class-cards">
-		<!--[if lt IE 7]>
-			<p class="browsehappy">
-				You are using an <strong>outdated</strong> browser. Please <a href="#">upgrade your browser</a> to improve your experience.
-			</p>
-		<![endif]-->
-		<!-- Nav Bar -->
-		<div>
-			<nav class="navbar navbar-expand-lg navbar-light bg-light">
-				<a class="navbar-brand" href="#">Comp-110</a>
-				<button
-					class="navbar-toggler"
-					type="button"
-					data-toggle="collapse"
-					data-target="#navbarNavAltMarkup"
-					aria-controls="navbarNavAltMarkup"
-					aria-expanded="false"
-					aria-label="Toggle navigation"
-				>
-					<span class="navbar-toggler-icon"></span>
-				</button>
-				<div class="collapse navbar-collapse" id="navbarNavAltMarkup">
-					<div class="navbar-nav">
-						<a class="nav-item nav-link" href="./index.html">DOM</a>
-						<a class="nav-item nav-link active" href="./index_cards.html">Cards<span class="sr-only">(current)</span></a>
-						<a class="nav-item nav-link disabled" href="#">State</a>
-						<a class="nav-item nav-link disabled" href="#">React</a>
-					</div>
-				</div>
-			</nav>
-		</div>
-		<div class="container container-cards">
-			<h1><strong>Working with Cards</strong></h1>
-		</div>
-		<main>
-			<div class="box-6">
-				<button id="btn-add-card">Add Card</button>
-			</div>
-			<div class="box-1" id="original">
-				<p class="content">Card-1</p>
-				<button class="content">Add Before</button>
-				<button class="content">Add After</button><br />
-				<button class="content">Delete</button>
-			</div>
-			<div class="box-2">
-				<p class="content">Card-2</p>
-				<button class="content">Add Before</button>
-				<button class="content">Add After</button><br />
-				<button class="content">Delete</button>
-			</div>
-			<div class="box-3">
-				<p class="content">Card-3</p>
-				<button class="content">Add Before</button>
-				<button class="content">Add After</button><br />
-				<button class="content">Delete</button>
-			</div>
-			<div class="box-4">
-				<p class="content">Card-4</p>
-				<button class="content">Add Before</button>
-				<button class="content">Add After</button><br />
-				<button class="content">Delete</button>
-			</div>
-			<div class="box-5">
-				<p class="content">Card-5</p>
-				<p class="content"><strong style="font-size: 2rem;">Right Side</strong></p>
-			</div>
-		</main>
-
-		<script src="./scripts/main_cards.js" type="module" async defer></script>
-		<!-- Optional JavaScript -->
-		<!-- jQuery first, then Popper.js, then Bootstrap JS -->
-		<script
-			src="https://code.jquery.com/jquery-3.2.1.slim.min.js"
-			integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN"
-			crossorigin="anonymous"
-		></script>
-		<script
-			src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.12.9/umd/popper.min.js"
-			integrity="sha384-ApNbgh9B+Y1QKtv3Rn7W3mgPxhU9K/ScQsAP7hUibX39j7fakFPskvXusvfa0b4Q"
-			crossorigin="anonymous"
-		></script>
-		<script
-			src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js"
-			integrity="sha384-JZR6Spejh4U02d8jOt6vLEHfe/JQGiRRSQQxSfFWpi1MquVdAyjUar5+76PVCmYl"
-			crossorigin="anonymous"
-		></script>
-	</body>
-</html>
