--- src/02-dom/sandbox/scripts/main_cards.js
@@ -1,5 +1,0 @@
-import functionsCards from "./functions_cards.js";
-
-/*
-add event listeners below
-*/
