--- src/02-dom/scripts/main.js
@@ -1,5 +1,0 @@
-import functions from "./functions.js";
-
-/*
-add event listeners below
-*/
